refactor: DownloadWorker 저장 및 알람 코드 분리
