+++ android/app/src/main/AndroidManifest.xml
@@ -99,7 +99,7 @@
             tools:node="merge" />
 
         <receiver
-            android:name="com.andlife.data.util.media.DownloadCancelReceiver"
+            android:name="com.andlife.data.util.media.download.DownloadCancelReceiver"
             android:exported="false" />
 
     </application>
